requirements.txt
python-telegram-bot~=12.4.2
pyTelegramBotAPI~=3.6.7
google-api-core~=1.22.1
google-api-python-client~=1.11.0
google-auth~=1.18.0
google-auth-httplib2>=0.0.3
google-auth-oauthlib>=0.4.1
pandas~=1.1.1
pyjarowinkler~=1.8
mock~=4.0.2
parse~=1.18.0
SQLAlchemy~=1.3.23
pytest>=6.2.2